refactor(home page): improve code and styles
