+++ src/components/header/header.css
@@ -9,8 +9,6 @@
 	flex-wrap: wrap;
 	justify-content: space-between;
 	align-items: center;
-
-	/* position: sticky; */
 	position: fixed;
 	left: 0;
 	right: 0;
