+++ src/helpers/durationFormatHelper.js
