+++ src/helpers/ratingFormatHelper.js
@@ -1,0 +1,3 @@
+export const formatRating = rating => {
+	return rating.toFixed(1)
+}
