+++ src/index.html
@@ -4,7 +4,7 @@
 		<meta charset="UTF-8" />
 		<meta http-equiv="X-UA-Compatible" content="IE=edge" />
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<title>Davai Film</title>
+		<title>DDFilms</title>
 		<link
 			rel="apple-touch-icon"
 			sizes="180x180"
