+++ src/mocks/films.js
@@ -6,7 +6,7 @@ export const TOPFILM = {
 	genre: 'Фантастика',
 	duration: 169,
 	desription:
-		'Продолжение эпической саги о Полу Атрейдесе. Он продолжает путь к тому, чтобы стать МуадДибом, в то время как его враги плетут заговоры против него.',
+		'Продолжение эпической саги о Поле Атрейдесе. Он продолжает путь к тому, чтобы стать МуадДибом, в то время как его враги плетут заговоры против него.',
 }
 
 export function getFilmsMocks(limit, offset = 0) {
